feat: price 값객체로 포장 및 값은 long타입으로 변환

+++ src/test/java/kitchenpos/application/fixture/MenuFixture.java
@@ -6,7 +6,6 @@ import static kitchenpos.application.fixture.ProductFixture.PRODUCT_ID_ONE;
 import static kitchenpos.application.fixture.ProductFixture.PRODUCT_ID_THREE;
 import static kitchenpos.application.fixture.ProductFixture.PRODUCT_ID_TWO;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import kitchenpos.domain.Menu;
@@ -17,9 +16,9 @@ public class MenuFixture {
     public static final Long MENU_FIRST_ID = 1L;
     public static final Long MENU_SECOND_ID = 2L;
     public static final Long MENU_THIRD_ID = 3L;
-    public static final BigDecimal MENU_PRICE = BigDecimal.valueOf(9900);
-    public static final BigDecimal MENU_INVALID_PRICE = BigDecimal.valueOf(-10);
-    public static final BigDecimal MENU_EXPENSIVE_PRICE = BigDecimal.valueOf(11001);
+    public static final long MENU_PRICE = 9900L;
+    public static final long MENU_INVALID_PRICE = -10L;
+    public static final long MENU_EXPENSIVE_PRICE = 11001L;
     public static final String MENU_NAME = "빵";
     public static final List<MenuProduct> MENU_PRODUCTS = new ArrayList<>();
     public static final List<MenuProduct> INVALID_MENU_PRODUCTS = new ArrayList<>();
