+++ src/test/java/kitchenpos/application/fixture/ProductFixture.java
@@ -1,6 +1,5 @@
 package kitchenpos.application.fixture;
 
-import java.math.BigDecimal;
 import kitchenpos.domain.Product;
 
 public class ProductFixture {
@@ -10,7 +9,7 @@ public class ProductFixture {
     public static final Long PRODUCT_ID_THREE = 3L;
     public static final Long INVALID_PRODUCT_ID = -1L;
     public static final String PRODUCT_NAME = "맛잇는 빵";
-    public static final BigDecimal PRODUCT_PRICE = BigDecimal.valueOf(1000);
-    public static final BigDecimal INVALID_PRODUCT_PRICE = BigDecimal.valueOf(-1);
+    public static final long PRODUCT_PRICE = 1000L;
+    public static final long INVALID_PRODUCT_PRICE = -1L;
     public static final Product UNSAVED_PRODUCT = new Product(PRODUCT_NAME, PRODUCT_PRICE);
 }
